refactor(checkbox): move from experimental to components
BREAKING CHANGE: Checkbox and CheckboxGroup have moved
to src/components/checkbox
