+++ src/components/accordion/accordion.stories.mdx
@@ -6,7 +6,7 @@ import { AccordionGroup, Accordion } from ".";
 import Textbox from "../textbox";
 import Button from "../button";
 import { Dl, Dt, Dd } from "../definition-list";
-import { Checkbox } from "../../__experimental__/components/checkbox";
+import { Checkbox } from "../checkbox";
 import Box from "../box";
 
 <Meta
