+++ src/components/checkbox/checkbox-validations.stories.mdx
@@ -4,7 +4,7 @@ import { Meta, Story, Preview, Props } from "@storybook/addon-docs/blocks";
 import { Checkbox, CheckboxGroup } from ".";
 
 <Meta
-  title="Experimental/Checkbox/Validations"
+  title="Design System/Checkbox/Validations"
   parameters={{ info: { disable: true } }}
 />
