+++ src/components/checkbox/checkbox.stories.mdx
@@ -1,10 +1,10 @@
 import { useState } from "react";
 import { Meta, Story, Preview, Props } from "@storybook/addon-docs/blocks";
-import StyledSystemProps from "../../../../.storybook/utils/styled-system-props";
+import StyledSystemProps from "../../../.storybook/utils/styled-system-props";
 
 import { Checkbox, CheckboxGroup } from ".";
 
-<Meta title="Experimental/Checkbox" parameters={{ info: { disable: true } }} />
+<Meta title="Design System/Checkbox" parameters={{ info: { disable: true } }} />
 
 # Checkbox
 
@@ -21,7 +21,7 @@ import { Checkbox, CheckboxGroup } from ".";
 ## Quick Start
 
 ```javascript
-import { Checkbox } from "carbon-react/lib/__experimental__/components/checkbox";
+import { Checkbox } from "carbon-react/lib/components/checkbox";
 ```
 
 ## Designer Notes
