+++ src/components/checkbox/index.js
