+++ src/components/dialog/dialog.stories.js
@@ -7,7 +7,7 @@ import Form from "../form";
 import Textbox from "../textbox";
 import Button from "../button";
 import DateInput from "../../__experimental__/components/date";
-import { Checkbox } from "../../__experimental__/components/checkbox";
+import { Checkbox } from "../checkbox";
 import { Select, Option } from "../select";
 
 export default {
