+++ src/components/drawer/drawer.stories.mdx
@@ -4,7 +4,7 @@ import { text, object } from "@storybook/addon-knobs";
 import { action } from "@storybook/addon-actions";
 import styled from "styled-components";
 
-import { Checkbox } from "../../__experimental__/components/checkbox";
+import { Checkbox } from "../checkbox";
 import Search from "../search";
 import Drawer from ".";
 import Button from "../button";
